comments, tidyups and seperate pdftextrecognition has regersion in matching newline
comments, tidyups and seperate pdftextrecognition has regersion in matching newline that was probably introduced when I implemented ale, if else simplification without properly checking it first.
